Fleshing out API

+++ models/manager.go
@@ -3,6 +3,6 @@ package models
 //Manager is also a user
 type Manager struct {
 	FullName string `json:"name"`
-	ID       int    `json:"id"`
-	TeamID   int    `json:"team_id"`
+	ID       int64  `json:"id"`
+	TeamID   int64  `json:"team_id"`
 }
